fix: trait result mockdata 세팅
