+++ src/app/(fullscreen)/trait/page.tsx
@@ -31,26 +31,9 @@ export default function Trait() {
     }
   };
 
-  const onSubmit = methods.handleSubmit(async (data) => {
-    const { mood, criterion, content } = data;
-
-    try {
-      const query = {
-        step1Question: mood,
-        step2Question: criterion,
-        favoriteCategory: content,
-      };
-
-      const res = await axios.post("/api/user-type", null, { params: query });
-
-      const { userTypeLabel, description } = res.data.result;
-
-      router.push(
-        `${PATHS.TRAIT_RESULT}?label=${encodeURIComponent(userTypeLabel)}&desc=${encodeURIComponent(description)}`,
-      );
-    } catch (err) {
-      console.error(err);
-    }
+  const onSubmit = methods.handleSubmit((data) => {
+    console.log("제출된 데이터:", data);
+    router.push(PATHS.TRAIT_RESULT);
   });
 
   const mood = methods.watch("mood");
